requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2.2
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
future==0.18.2
httpie==2.4.0
idna==2.10
Markdown==3.3.4
Pillow==8.2.0
Pygments==2.9.0
PySocks==1.7.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4